requirements.txt
faiss-gpu
git+https://github.com/facebookresearch/pytorch3d.git
git+https://github.com/JonathonLuiten/diff-gaussian-rasterization-w-depth.git@cb65e4b86bc3bd8ed42174b72a62e8d3a3a71110
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
git+https://github.com/tfy14esa/evaluate_3d_reconstruction_lib.git
imageio
lpips
mathutils
matplotlib
natsort
open3d==0.16.0
opencv-python
pandas
pytorch-msssim
pyyaml
rtree==0.9.7
scikit-image
scikit-learn
scipy
seaborn
torchmetrics
tqdm
trimesh==3.21.5
tyro
